+++ src/api/api.js
@@ -70,7 +70,7 @@ class Api {
 
   sync(data) {
     return this._load({
-      url: 'tasks/sync',
+      url: 'points/sync',
       method: Method.POST,
       body: JSON.stringify(data),
       headers: new Headers({'Content-Type': 'application/json'}),
