+++ src/view/statistics.js
@@ -1,243 +1,7 @@
 import SmartView from './smart.js';
-import Chart from 'chart.js';
-import ChartDataLabels from 'chartjs-plugin-datalabels';
-import {msToTime} from '../utils/common.js';
+import {renderTypesChart, renderMoneysChart, renderDurationChart} from '../utils/statistics';
 
-const makeItemsUniq = (items) => [...new Set(items)];
-const countEventsByType = (events, type) => {
-  return events.filter((event) => event.type === type).length;
-};
-const countPriceByType = (events, type) => {
-  const eventsByType = events.filter((event) => event.type === type);
-  let totalPrice = 0;
-  eventsByType.forEach((item) => totalPrice += item.basePrice);
-  return totalPrice;
-};
-const countDurationByType = (events, type) => {
-  const eventsByType = events.filter((event) => event.type === type);
-  let duration = 0;
-  eventsByType.forEach((item) => duration += item.dateTo - item.dateFrom);
-  return duration;
-};
-
-const renderTypesChart = (typeCtx, events) => {
-  const types = events.map((event) => event.type);
-  const uniqTypes = makeItemsUniq(types);
-  const eventsByTypeCounts = uniqTypes.map((item) => countEventsByType(events, item));
-  const uniqTypesUpperCase = uniqTypes.map((item) => item.toUpperCase());
-
-  const typeChart = new Chart(typeCtx, {
-    plugins: [ChartDataLabels],
-    type: 'horizontalBar',
-    data: {
-      labels: uniqTypesUpperCase,
-      datasets: [{
-        data: eventsByTypeCounts,
-        backgroundColor: '#ffffff',
-        hoverBackgroundColor: '#ffffff',
-        anchor: 'start',
-        barThickness: 44,
-        minBarLength: 50,
-      }],
-    },
-    options: {
-      plugins: {
-        datalabels: {
-          font: {
-            size: 13,
-          },
-          color: '#000000',
-          anchor: 'end',
-          align: 'start',
-          formatter: (val) => `${val}x`,
-        },
-      },
-      title: {
-        display: true,
-        text: 'TYPE',
-        fontColor: '#000000',
-        fontSize: 23,
-        position: 'left',
-      },
-      scales: {
-        yAxes: [{
-          ticks: {
-            fontColor: '#000000',
-            padding: 5,
-            fontSize: 13,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-        xAxes: [{
-          ticks: {
-            display: false,
-            beginAtZero: true,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-      },
-      legend: {
-        display: false,
-      },
-      tooltips: {
-        enabled: false,
-      },
-    },
-  });
-  return typeChart;
-};
-const renderMoneysChart = (moneyCtx, events) => {
-  const types = events.map((event) => event.type);
-  const uniqTypes = makeItemsUniq(types);
-  const eventsByPriceCounts = uniqTypes.map((item) => countPriceByType(events, item));
-  const uniqTypesUpperCase = uniqTypes.map((item) => item.toUpperCase());
-
-  const moneyChart = new Chart(moneyCtx, {
-    plugins: [ChartDataLabels],
-    type: 'horizontalBar',
-    data: {
-      labels: uniqTypesUpperCase,
-      datasets: [{
-        data: eventsByPriceCounts,
-        backgroundColor: '#ffffff',
-        hoverBackgroundColor: '#ffffff',
-        anchor: 'start',
-        barThickness: 44,
-        minBarLength: 50,
-      }],
-    },
-    options: {
-      plugins: {
-        datalabels: {
-          font: {
-            size: 13,
-          },
-          color: '#000000',
-          anchor: 'end',
-          align: 'start',
-          formatter: (val) => `€ ${val}`,
-        },
-      },
-      title: {
-        display: true,
-        text: 'MONEY',
-        fontColor: '#000000',
-        fontSize: 23,
-        position: 'left',
-      },
-      scales: {
-        yAxes: [{
-          ticks: {
-            fontColor: '#000000',
-            padding: 5,
-            fontSize: 13,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-        xAxes: [{
-          ticks: {
-            display: false,
-            beginAtZero: true,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-      },
-      legend: {
-        display: false,
-      },
-      tooltips: {
-        enabled: false,
-      },
-    },
-  });
-
-  return moneyChart;
-};
-const renderDurationChart = (timeCtx, events) => {
-  const types = events.map((event) => event.type);
-  const uniqTypes = makeItemsUniq(types);
-  const eventsByDurationCounts = uniqTypes.map((item) => countDurationByType(events, item));
-  const uniqTypesUpperCase = uniqTypes.map((item) => item.toUpperCase());
-
-  const timeChart = new Chart(timeCtx, {
-    plugins: [ChartDataLabels],
-    type: 'horizontalBar',
-    data: {
-      labels: uniqTypesUpperCase,
-      datasets: [{
-        data: eventsByDurationCounts,
-        backgroundColor: '#ffffff',
-        hoverBackgroundColor: '#ffffff',
-        anchor: 'start',
-        barThickness: 44,
-        minBarLength: 50,
-      }],
-    },
-    options: {
-      plugins: {
-        datalabels: {
-          font: {
-            size: 13,
-          },
-          color: '#000000',
-          anchor: 'end',
-          align: 'start',
-          formatter: (val) => `${msToTime(val)}`,
-        },
-      },
-      title: {
-        display: true,
-        text: 'TIME-SPEND',
-        fontColor: '#000000',
-        fontSize: 23,
-        position: 'left',
-      },
-      scales: {
-        yAxes: [{
-          ticks: {
-            fontColor: '#000000',
-            padding: 5,
-            fontSize: 13,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-        xAxes: [{
-          ticks: {
-            display: false,
-            beginAtZero: true,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-      },
-      legend: {
-        display: false,
-      },
-      tooltips: {
-        enabled: false,
-      },
-    },
-  });
-
-  return timeChart;
-};
+const BAR_HEIGHT = 55;
 
 const createStatisticsTemplate = () => {
   return `<section class="statistics">
@@ -292,7 +56,6 @@ class Statistics extends SmartView {
     const typeCtx = this.getElement().querySelector('.statistics__chart--transport');
     const timeCtx = this.getElement().querySelector('.statistics__chart--time');
 
-    const BAR_HEIGHT = 55;
     moneyCtx.height = BAR_HEIGHT * 7.5;
     typeCtx.height = BAR_HEIGHT * 7.5;
     timeCtx.height = BAR_HEIGHT * 7.5;
